+++ experiment/measurer/test_data/test_measure_snapshot_coverage/corpus-archive-0001.tar.gz
