+++ experiment/measurer/test_data/test_measure_snapshot_coverage/freetype2-2017-coverage
